Trait ckb_rocksdb::ops::GetPinned

pub trait GetPinned<'a> {
    type ReadOptions;
    fn get_pinned_full<K: AsRef<[u8]>>(
        &'a self, 
        key: K, 
        readopts: Option<Self::ReadOptions>
    ) -> Result<Option<DBPinnableSlice<'a>>, Error>;

    fn get_pinned<K: AsRef<[u8]>>(
        &'a self, 
        key: K
    ) -> Result<Option<DBPinnableSlice<'a>>, Error> { ... }

    fn get_pinned_opt<K: AsRef<[u8]>>(
        &'a self, 
        key: K, 
        readopts: Self::ReadOptions
    ) -> Result<Option<DBPinnableSlice<'a>>, Error> { ... }
}

Associated Types

type ReadOptions

Loading content...
Required methods

fn get_pinned_full<K: AsRef<[u8]>>(
&'a self,
key: K,
readopts: Option<Self::ReadOptions>
) -> Result<Option<DBPinnableSlice<'a>>, Error>

Return the value associated with a key using RocksDB's PinnableSlice so as to avoid unnecessary memory copy.

Loading content...
Provided methods

fn get_pinned<K: AsRef<[u8]>>(
&'a self,
key: K
) -> Result<Option<DBPinnableSlice<'a>>, Error>

Return the value associated with a key using RocksDB's PinnableSlice so as to avoid unnecessary memory copy. Similar to get_pinned_opt but leverages default options.

fn get_pinned_opt<K: AsRef<[u8]>>(
&'a self,
key: K,
readopts: Self::ReadOptions
) -> Result<Option<DBPinnableSlice<'a>>, Error>

Return the value associated with a key using RocksDB's PinnableSlice so as to avoid unnecessary memory copy.

Loading content...