Trait ckb_rocksdb::ops::GetPinnedCF

pub trait GetPinnedCF<'a> {
    type ColumnFamily;
    type ReadOptions;
    fn get_pinned_cf_full<K: AsRef<[u8]>>(
        &'a self, 
        cf: Option<Self::ColumnFamily>, 
        key: K, 
        readopts: Option<Self::ReadOptions>
    ) -> Result<Option<DBPinnableSlice<'a>>, Error>;

    fn get_pinned_cf<K: AsRef<[u8]>>(
        &'a self, 
        cf: Self::ColumnFamily, 
        key: K
    ) -> Result<Option<DBPinnableSlice<'a>>, Error> { ... }

    fn get_pinned_cf_opt<K: AsRef<[u8]>>(
        &'a self, 
        cf: Self::ColumnFamily, 
        key: K, 
        readopts: Self::ReadOptions
    ) -> Result<Option<DBPinnableSlice<'a>>, Error> { ... }
}

Associated Types

type ColumnFamily

type ReadOptions

Loading content...
Required methods

fn get_pinned_cf_full<K: AsRef<[u8]>>(
&'a self,
cf: Option<Self::ColumnFamily>,
key: K,
readopts: Option<Self::ReadOptions>
) -> Result<Option<DBPinnableSlice<'a>>, Error>

Return the value associated with a key using RocksDB's PinnableSlice so as to avoid unnecessary memory copy.

Loading content...
Provided methods

fn get_pinned_cf<K: AsRef<[u8]>>(
&'a self,
cf: Self::ColumnFamily,
key: K
) -> Result<Option<DBPinnableSlice<'a>>, Error>

Return the value associated with a key using RocksDB's PinnableSlice so as to avoid unnecessary memory copy.

fn get_pinned_cf_opt<K: AsRef<[u8]>>(
&'a self,
cf: Self::ColumnFamily,
key: K,
readopts: Self::ReadOptions
) -> Result<Option<DBPinnableSlice<'a>>, Error>

Return the value associated with a key using RocksDB's PinnableSlice so as to avoid unnecessary memory copy.

Loading content...