Trait ckb_rocksdb::ops::GetPropertyCF

pub trait GetPropertyCF {
    fn property_value_cf(
        &self, 
        cf: &ColumnFamily, 
        name: &str
    ) -> Result<Option<String>, Error>;

    fn property_int_value_cf(
        &self, 
        cf: &ColumnFamily, 
        name: &str
    ) -> Result<Option<u64>, Error> { ... }
}

Required methods

fn property_value_cf(
&self,
cf: &ColumnFamily,
name: &str
) -> Result<Option<String>, Error>

Retrieves a RocksDB property by name, for a specific column family.

For a full list of properties, see https://github.com/facebook/rocksdb/blob/08809f5e6cd9cc4bc3958dd4d59457ae78c76660/include/rocksdb/db.h#L428-L634

Loading content...
Provided methods

fn property_int_value_cf(
&self,
cf: &ColumnFamily,
name: &str
) -> Result<Option<u64>, Error>

Retrieves a RocksDB property for a specific column family and casts it to an integer.

For a full list of properties that return int values, see https://github.com/facebook/rocksdb/blob/08809f5e6cd9cc4bc3958dd4d59457ae78c76660/include/rocksdb/db.h#L654-L689

Loading content...