Trait ckb_rocksdb::ops::OpenCF

pub trait OpenCF: OpenRaw {
    fn open_cf<P, I, N>(opts: &Options, path: P, cfs: I) -> Result<Self, Error>
    where
        P: AsRef<Path>,
        I: IntoIterator<Item = N>,
        N: AsRef<str>,
    { ... }

    fn open_cf_descriptors<P, I>(
        opts: &Options, 
        path: P, 
        cfs: I
    ) -> Result<Self, Error>
    where
        P: AsRef<Path>,
        I: IntoIterator<Item = ColumnFamilyDescriptor>,
    { ... }

    fn open_cf_descriptors_with_descriptor<P, I>(
        opts: &Options, 
        path: P, 
        cfs: I, 
        descriptor: Self::Descriptor
    ) -> Result<Self, Error>
    where
        P: AsRef<Path>,
        I: IntoIterator<Item = ColumnFamilyDescriptor>,
    { ... }
}

Provided methods

fn open_cf<P, I, N>(opts: &Options, path: P, cfs: I) -> Result<Self, Error> where
P: AsRef<Path>,
I: IntoIterator<Item = N>,
N: AsRef<str>,

Open a database with the given database options and column family names.

Column families opened using this function will be created with default Options.

fn open_cf_descriptors<P, I>(
opts: &Options,
path: P,
cfs: I
) -> Result<Self, Error> where
P: AsRef<Path>,
I: IntoIterator<Item = ColumnFamilyDescriptor>,

Open a database with the given database options and column family descriptors.

fn open_cf_descriptors_with_descriptor<P, I>(
opts: &Options,
path: P,
cfs: I,
descriptor: Self::Descriptor
) -> Result<Self, Error> where
P: AsRef<Path>,
I: IntoIterator<Item = ColumnFamilyDescriptor>,

Loading content...