Struct ckb_rocksdb::prelude::ReadOptions

pub struct ReadOptions { /* fields omitted */ }

Methods

impl ReadOptions

pub fn set_snapshot<T>(&mut self, snapshot: &T) where
T: ConstHandle<rocksdb_snapshot_t>,

pub fn set_iterate_upper_bound<K: AsRef<[u8]>>(&mut self, key: K)

pub fn set_prefix_same_as_start(&mut self, v: bool)

pub fn set_total_order_seek(&mut self, v: bool)

pub fn set_readahead_size(&mut self, v: usize)

If non-zero, an iterator will create a new table reader which performs reads of the given size. Using a large size (> 2MB) can improve the performance of forward iteration on spinning disks. Default: 0

use ckb_rocksdb::{ReadOptions};

let mut opts = ReadOptions::default();
opts.set_readahead_size(4_194_304); // 4mb

pub fn input_or_default(
input: Option<&ReadOptions>,
default_readopts: &mut Option<ReadOptions>
) -> Result<*mut rocksdb_readoptions_t, Error>