Struct ckb_rocksdb::WriteBatch

pub struct WriteBatch { /* fields omitted */ }

An atomic batch of write operations.

Making an atomic commit of several writes:

use ckb_rocksdb::{prelude::*, WriteBatch};

let path = "_path_for_rocksdb_storage1";

let db = DB::open_default(&path).unwrap();

let mut batch = WriteBatch::default();
batch.put(b"my key", b"my value");
batch.put(b"key2", b"value2");
batch.put(b"key3", b"value3");

db.write(batch); // Atomically commits the batch

Methods

impl WriteBatch

pub fn len(&self) -> usize

pub fn size_in_bytes(&self) -> usize

Return WriteBatch serialized size (in bytes).

pub fn is_empty(&self) -> bool

pub fn put<K, V>(&mut self, key: K, value: V) -> Result<(), Error> where
K: AsRef<[u8]>,
V: AsRef<[u8]>,

Insert a value into the database under the given key.

pub fn put_cf<K, V>(
&mut self,
cf: &ColumnFamily,
key: K,
value: V
) -> Result<(), Error> where
K: AsRef<[u8]>,
V: AsRef<[u8]>,

pub fn merge<K, V>(&mut self, key: K, value: V) -> Result<(), Error> where
K: AsRef<[u8]>,
V: AsRef<[u8]>,

pub fn merge_cf<K, V>(
&mut self,
cf: &ColumnFamily,
key: K,
value: V
) -> Result<(), Error> where
K: AsRef<[u8]>,
V: AsRef<[u8]>,

pub fn delete<K: AsRef<[u8]>>(&mut self, key: K) -> Result<(), Error>

Remove the database entry for key.

Returns an error if the key was not found.

pub fn delete_cf<K: AsRef<[u8]>>(
&mut self,
cf: &ColumnFamily,
key: K
) -> Result<(), Error>

pub fn delete_range<K: AsRef<[u8]>>(
&mut self,
from: K,
to: K
) -> Result<(), Error>

Remove database entries from start key to end key.

Removes the database entries in the range ["begin_key", "end_key"), i.e., including "begin_key" and excluding "end_key". It is not an error if no keys exist in the range ["begin_key", "end_key").

pub fn delete_range_cf<K: AsRef<[u8]>>(
&mut self,
cf: ColumnFamily,
from: K,
to: K
) -> Result<(), Error>

Remove database entries in column family from start key to end key.

Removes the database entries in the range ["begin_key", "end_key"), i.e., including "begin_key" and excluding "end_key". It is not an error if no keys exist in the range ["begin_key", "end_key").

pub fn clear(&mut self) -> Result<(), Error>

Clear all updates buffered in this batch.